refactor: Extract EmojiReactionNotAllowedReason outside of Message and into its own file
