+++ app/src/main/java/com/infomaniak/mail/utils/ErrorCode.kt
@@ -19,7 +19,7 @@ package com.infomaniak.mail.utils
 
 import com.infomaniak.lib.core.utils.ApiErrorCode
 import com.infomaniak.mail.R
-import com.infomaniak.mail.data.models.message.Message.EmojiReactionNotAllowedReason
+import com.infomaniak.mail.data.models.message.EmojiReactionNotAllowedReason
 import com.infomaniak.lib.core.R as RCore
 
 @Suppress("MemberVisibilityCanBePrivate")
